requirements.txt
streamlit
torch>=2.0.0
timm
Pillow
torchvision
pandas
numpy
scikit-learn